fix: 完善获取合格率排行TOP3的接口

+++ project/src/main/java/com/cby/tcs/controller/JobQualifiedPercentController.java
@@ -61,7 +61,7 @@ public class JobQualifiedPercentController {
    */
   @GetMapping("/passRateRanking")
   public ResultEntity passRateRanking(@RequestParam JobType type) {
-    List<Option<Double>> options = jobQualifiedPercentService.passRateRanking(type);
+    List<Option<Float>> options = jobQualifiedPercentService.passRateRanking(type);
     return ResultEntity.success(options);
   }
